Added theoreical scale detection
added functions to ScaleGenerator that detects whether the selected scale is theoretical or not.  Added text to the scenes that will activate if the scale is theoretical.
The warning text only shows in the ScaleScene.  Need to redesign the other scenes so the warning does not clutter the screen too much.