ScaleScene now saves and updates the dropdown values properly
The bug was caused by a logic error.  The OnValueChanged Event was triggered any time the scaleDrop or rootDrop values were changed.  This lead to the UpdateScale() function to be called before rootDrop/scaleDrop values were updated to the PlayerPrefs

To fix this, I added a bool variable that is false on Awake(), and then changes to true after the scaleDrop and rootDrop have been updated by the PlayerPrefs.

The UpdateScale() function automatically returns if the bool is false.